feat(router): Validate tree thresholds are correct, auto-fix and warn
Given the hierchical structure of the router nodes, we cannot allow a
parent threshold to be higher than any children threshold, because that
would lead to earlier tree-branch discarding.

So, now, we ensure that nodes have, as maximum, the min children
threshold. It's auto-fixed on loading (not modifying the router.yml
file) and warned in logs (to ensure they are manually fixed in the
file).